service_updater/requirements.txt
annotated-types==0.5.0
et-xmlfile==1.1.0
greenlet==2.0.2
openpyxl==3.1.2
psycopg2==2.9.7
pydantic==1.9.0
pydantic_core==2.6.3
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
SQLAlchemy==1.4.36
typing_extensions==4.7.1
xlrd==2.0.1